+++ bootstrap/test/Spec.hs
@@ -3,6 +3,7 @@ import Test.Tasty
 import Test.Driver.Transform.GraphAndType (transformationShouldKeepTypeIntact)
 import Test.Driver.Unification (unifyingOfUnifiableTypesShouldNotGoWrong)
 import Test.Inference (inferringExpressionWithNoError, inferringExpressionWithError)
+import Test.Graph.Data (test'GraphData)
 
 main :: IO ()
 main = defaultMain globalTests
@@ -12,4 +13,7 @@ globalTests = testGroup "Global Test"
   [ testGroup "Transformation Test" [transformationShouldKeepTypeIntact]
   , testGroup "Graph Unification Test" [unifyingOfUnifiableTypesShouldNotGoWrong]
   , testGroup "Inference Test" [inferringExpressionWithNoError, inferringExpressionWithError]
+  , testGroup "Algebraic graph" [
+    test'GraphData
+  ]
   ]